+++ api/src/main/scala/dx/api/DxFile.scala
@@ -3,10 +3,9 @@ package dx.api
 import dx.AppInternalException
 import dx.api.DxPath.DxScheme
 import spray.json._
-import dx.util.Enum
+import dx.util.{Enum, FileUtils}
 
 import java.net.URI
-import java.nio.file.Paths
 
 case class DxFilePart(state: String, size: Long, md5: String)
 
@@ -149,7 +148,7 @@ object DxFile {
     */
   def format(fileId: String, folder: String, name: String, project: Option[String]): String = {
     val authority = project.map(proj => s"${proj}:${fileId}::").getOrElse(s"${fileId}::")
-    val path = Paths.get(folder).resolve(name).toString
+    val path = FileUtils.getPath(folder).resolve(name).toString
     new URI(DxScheme, authority, path, null, null).toString
   }