+++ common/src/main/scala/dx/util/FileSource.scala
@@ -316,24 +316,25 @@ case class LocalFileSource(
     val parent = if (isDirectory) this else getParent.get
     val parentPath = parent.canonicalPath
     val newPath = parentPath.resolve(path)
+    val newCanonicalPath = FileUtils.normalizePath(newPath)
     cachedListing
       .flatMap { listing =>
         listing.collectFirst {
-          case fs: LocalFileSource if fs.canonicalPath == newPath => fs
+          case fs: LocalFileSource if fs.canonicalPath == newCanonicalPath => fs
         }
       }
       .getOrElse {
-        val newIsDirectory = newPath.toFile match {
+        val newIsDirectory = newCanonicalPath.toFile match {
           case f if f.exists()         => f.isDirectory
           case _ if path.endsWith("/") => true
           case _                       => false
         }
         // the parent can only be used if it is the direct ancestor of the new path
-        val cachedParent = if (newPath.getParent == parentPath) Some(parent) else None
-        LocalFileSource(newPath, encoding, newIsDirectory)(newPath.toString,
-                                                           newPath,
-                                                           cachedParent = cachedParent,
-                                                           logger = logger)
+        val cachedParent = if (newCanonicalPath.getParent == parentPath) Some(parent) else None
+        LocalFileSource(newCanonicalPath, encoding, newIsDirectory)(newPath.toString,
+                                                                    newPath,
+                                                                    cachedParent = cachedParent,
+                                                                    logger = logger)
       }
   }
 
@@ -481,7 +482,7 @@ object LocalFileSource {
     if (Files.exists(path)) {
       path.toRealPath()
     } else if (path.isAbsolute) {
-      path
+      FileUtils.normalizePath(path)
     } else {
       findInPath(path.toString, searchPath).getOrElse(
           // it's a non-existant relative path - localize it to current working dir
@@ -536,7 +537,7 @@ case class HttpFileSource(
 )(override val address: String)
     extends AbstractAddressableFileNode(address, encoding) {
 
-  private lazy val path = Paths.get(uri.getPath)
+  private lazy val path = FileUtils.getPath(uri.getPath)
 
   override lazy val name: String = path.getFileName.toString
 
@@ -751,7 +752,7 @@ case class FileSourceResolver(protocols: Vector[FileAccessProtocol]) {
             // the imported file is not relative to the parent, but
             // but LocalFileAccessProtocol may be configured to look
             // for it in a different folder
-            fromFile(Paths.get(address))
+            fromFile(FileUtils.getPath(address))
           case other =>
             throw new Exception(s"Not an AddressableFileNode: ${other}")
         }
@@ -778,7 +779,7 @@ case class FileSourceResolver(protocols: Vector[FileAccessProtocol]) {
             // the imported file is not relative to the parent, but
             // but LocalFileAccessProtocol may be configured to look
             // for it in a different folder
-            fromFile(Paths.get(address))
+            fromFile(FileUtils.getPath(address))
           case other =>
             throw new Exception(s"Not an AddressableFileNode: ${other}")
         }