+++ protocols/RELEASE_NOTES.md
@@ -2,6 +2,10 @@
 
 ## in develop
 
+* Normalizes any file/folder paths
+
+## 0.4.0 (2021-06-23)
+
 * Exposes dx folder in `DxFolderSource`
 * Implements recursive listing in `S3FileAccessProtocol` and `DxFileAccessProtocol`